python-developer/19-graphql/szkolenie-02-strawberry/requirements.txt
django
strawberry-graphql-django
